Changed cart_create_indexed_particles default particle precision to double Added makefile

+++ utils/cart_create_indexed_particles/cart_create_indexed_particles.c
@@ -12,7 +12,7 @@
 #define max(x,y)        (((x) > (y)) ? (x): (y))
 
 #define MAX_PARTICLE_SPECIES	10
-typedef float particle_float;
+typedef double particle_float;
 
 typedef struct {
 	float aexpn;
